+++ arithmetization/src/main/java/net/consensys/linea/zktracer/module/hub/fragment/imc/oob/precompiles/common/postCancun/fixedSizeFixedGasCost/P256VerifyOobCall.java
@@ -13,44 +13,44 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package net.consensys.linea.zktracer.module.hub.fragment.imc.oob.precompiles.common.bls.fixedSizeFixedGasCost;
+package net.consensys.linea.zktracer.module.hub.fragment.imc.oob.precompiles.common.postCancun.fixedSizeFixedGasCost;
 
-import static net.consensys.linea.zktracer.Trace.GAS_CONST_BLS_MAP_FP_TO_G1;
-import static net.consensys.linea.zktracer.Trace.OOB_INST_BLS_MAP_FP_TO_G1;
-import static net.consensys.linea.zktracer.Trace.PRECOMPILE_CALL_DATA_SIZE___FP_TO_G1;
-import static net.consensys.linea.zktracer.TraceCancun.Oob.CT_MAX_BLS_MAP_FP_TO_G1;
+import static net.consensys.linea.zktracer.Trace.GAS_CONST_P256_VERIFY;
+import static net.consensys.linea.zktracer.Trace.OOB_INST_P256_VERIFY;
+import static net.consensys.linea.zktracer.Trace.PRECOMPILE_CALL_DATA_SIZE___P256_VERIFY;
+import static net.consensys.linea.zktracer.TraceOsaka.Oob.CT_MAX_P256_VERIFY;
 
 import java.math.BigInteger;
 
 import net.consensys.linea.zktracer.Trace;
 
-public class BlsMapFpToG1OobCall extends BlsFixedSizeFixedGasCostOobCall {
-  public BlsMapFpToG1OobCall(BigInteger calleeGas) {
-    super(calleeGas, OOB_INST_BLS_MAP_FP_TO_G1);
+public class P256VerifyOobCall extends BlsFixedSizeFixedGasCostOobCall {
+  public P256VerifyOobCall(BigInteger calleeGas) {
+    super(calleeGas, OOB_INST_P256_VERIFY);
   }
 
   @Override
   long precompileExpectedCds() {
-    return PRECOMPILE_CALL_DATA_SIZE___FP_TO_G1;
+    return PRECOMPILE_CALL_DATA_SIZE___P256_VERIFY;
   }
 
   @Override
   long precompileLongCost() {
-    return GAS_CONST_BLS_MAP_FP_TO_G1;
+    return GAS_CONST_P256_VERIFY;
   }
 
   @Override
   protected void traceOobInstructionInOob(Trace.Oob trace) {
-    trace.isBlsMapFpToG1(true).oobInst(OOB_INST_BLS_MAP_FP_TO_G1);
+    trace.isP256Verify(true).oobInst(OOB_INST_P256_VERIFY);
   }
 
   @Override
   protected void traceOobInstructionInHub(Trace.Hub trace) {
-    trace.pMiscOobInst(OOB_INST_BLS_MAP_FP_TO_G1);
+    trace.pMiscOobInst(OOB_INST_P256_VERIFY);
   }
 
   @Override
   public int ctMax() {
-    return CT_MAX_BLS_MAP_FP_TO_G1;
+    return CT_MAX_P256_VERIFY;
   }
 }
